requirements.txt
APScheduler==3.8.0
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
colorzero==2.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.2.8
django-heroku==0.3.1
djangorestframework==3.12.4
drf-yasg==1.20.0
gpiozero==1.6.2
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
numpy==1.16.2
packaging==21.2
Pillow==8.3.2
psycopg2==2.9.1
pyparsing==2.4.7
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
sense-emu==1.1
six==1.16.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
tzlocal==2.1
uritemplate==4.1.1
urllib3==1.26.7
whitenoise==5.3.0
